refactor: adjust bid_methods table

+++ src/main/java/com/swp391_09/Koi_BE/repositories/BidMethodRepository.java
@@ -3,6 +3,6 @@ package com.swp391_09.Koi_BE.repositories;
 import com.swp391_09.Koi_BE.models.BidMethod;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BidMethodRepository extends JpaRepository<BidMethod, Long> {
+public interface BidMethodRepository extends JpaRepository<BidMethod, Integer> {
 
 }
